2. Model Deployment/requirements.txt
Flask
flask_wtf
flask-csp
jsonschema
requests
flask_cors
flask_limiter
python-dotenv
brotli
bcrypt
flask-talisman
email-validator
flask-session
pyotp
qrcode
pillow
python-dotenv